requirements.txt
streamlit
pandas
numpy
scikit-learn
sentence-transformers
pillow
plotly
